refactor(frontend): split BookListenPage into focused sub-components (#351)
Pure refactor — behaviour unchanged. The 693-line `BookListenPage`
component body is decomposed into focused sub-components under a new
`pages/listen/` directory, mirroring the existing `pages/metadata_edit/`
split pattern. No external contract changes (Playwright testids, JS
interop surface, `/api/rpc/progress` shape and the
`window.OmnibusAudio` control surface are all preserved).

Layout:
- `pages/listen.rs` (145 lines, was 812): thin orchestrator — fetches
  book metadata, owns playback signals, calls `bootstrap::install_audio_bootstrap`
  on web, dispatches into `ReadyPlayer` once loaded.
- `pages/listen/helpers.rs`: `format_hms`, `post_audio_progress`,
  `RATE_STEPS`, `audio_call`, `HLS_JS` asset + the relocated `format_hms`
  unit tests.
- `pages/listen/overlays.rs`: `FailedOverlay` + `PreparingOverlay`.
- `pages/listen/controls.rs`: `TopBar`, `AudioElement`, `Scrubber`,
  `TransportButtons`.
- `pages/listen/stage.rs`: `PlayerStage` (two-column cover + panel).
- `pages/listen/ready_player.rs`: `ReadyPlayer` post-load chrome +
  handler wiring.
- `pages/listen/bootstrap.rs`: web-only JS interop — `install_audio_bootstrap`
  composes `register_js_callbacks`, `inject_hls_script`,
  `install_control_surface`, and `run_manifest_init`.

Component body sizes (all under the ~150-line cap from the issue):
- `BookListenPage`: 97 lines
- `ReadyPlayer`: 93 lines
- `PlayerStage`: 40 lines
- everything else: <40 lines

Verification:
- `cargo fmt`
- `cargo clippy -p omnibus-frontend --features server --all-targets -- -D warnings` clean
- `cargo test -p omnibus-frontend --features server`: 108 passed (incl. the
  three relocated `format_hms` tests)
- `cargo check -p omnibus-frontend --features web` clean (pre-existing
  unrelated deprecation warning in `data/authors.rs`)
- `cargo check -p omnibus-frontend --no-default-features --features mobile` clean

Closes #351
